A-C funguje

+++ kostka/kostka_vek.py
@@ -7,7 +7,7 @@ from typing import Any
 KostkaVek = np.ndarray
 
 def nova_kostka_vek(n: int) -> KostkaVek:
-    return np.stack([ko.SLOZENA_KOSTKA]*n)
+    return np.stack([ko.SLOZENA_KOSTKA]*n, dtype=ko.SLOZENA_KOSTKA.dtype)
 
 def je_stejna(kostka1: KostkaVek, kostka2: KostkaVek) -> np.ndarray:
     return np.all(kostka1 == kostka2, axis=(1,2,3))
